+++ src/nl/b3p/viewer/config/services/ArcIMSService.java
@@ -256,7 +256,7 @@ public class ArcIMSService extends GeoService implements Updatable {
             updateLayers(update, linkedFS, result);
             removeOrphanLayersAfterUpdate(result);
             
-            if(linkedFS.getFeatureTypes().isEmpty()) {
+            if(linkedFS != null && linkedFS.getFeatureTypes().isEmpty()) {
                 log.debug("Linked ArcGISFeatureSource has no type names anymore, removing it");
                 Stripersist.getEntityManager().remove(linkedFS);
             }
